Added registration for event handler
